requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
discord==1.0.1
discord.py==1.5.1
idna==2.10
multidict==4.7.6
oauthlib==3.1.0
PySocks==1.7.1
python-dotenv==0.15.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.10.0
urllib3==1.26.2
yarl==1.5.1
